Converting Random Search algorithm output into Schedule Object
