+++ api/main.py
@@ -1,10 +1,11 @@
 """Algorithm 1 API Application"""
 
+from email.quoprimime import body_check
 from fastapi import FastAPI
 
 from . import dummy
 from .models import Schedule, ScheduleConstraints
-from .processInput import parseInput
+from .generate import generateSchedule
 
 app = FastAPI()
 
@@ -13,10 +14,7 @@ app = FastAPI()
 def post_schedule(body: ScheduleConstraints) -> Schedule:
     """Generates a schedule"""
 
-    parseInput(body)
-
-
-    return None
+    return generateSchedule(body)
 
 
 @app.post("/check_schedule")
